fix(rules): recognize mounted early returns
Reuse the client-only early-return proof across SSR diagnostics so post-hydration browser reads stay silent.
