+++ packages/oxlint-plugin-react-doctor/src/plugin/rules/performance/no-hydration-branch-on-browser-global.ts
@@ -10,6 +10,7 @@ import { hasDirective } from "../../utils/has-directive.js";
 import { hasEmailTemplateImport } from "../../utils/has-email-template-import.js";
 import { hasSuppressHydrationWarningAttribute } from "../../utils/has-suppress-hydration-warning-attribute.js";
 import { isAstNode } from "../../utils/is-ast-node.js";
+import { isAfterClientOnlyEarlyReturn } from "../../utils/is-after-client-only-early-return.js";
 import { isFunctionLike } from "../../utils/is-function-like.js";
 import { isGatedByFalsyInitialState } from "../../utils/is-gated-by-falsy-initial-state.js";
 import { isGeneratedImageRenderContext } from "../../utils/is-generated-image-render-context.js";
@@ -308,6 +309,7 @@ export const noHydrationBranchOnBrowserGlobal = defineRule({
       ) {
         return;
       }
+      if (isAfterClientOnlyEarlyReturn(predicateNode, componentOrHookNode)) return;
       const openingElement = findEnclosingJsxOpeningElement(predicateNode);
       if (hasSuppressHydrationWarningAttribute(openingElement)) return;
       if (branchHasSuppression(leftBranch) || (rightBranch && branchHasSuppression(rightBranch)))
